# rag/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CustomerServiceAgent RAG API",
    "description": "检索增强生成（RAG）系统后端接口。覆盖知识库文档管理、双模检索（向量 + 关键词）+ 融合去重 + Reranker 精排、以及检索增强问答（RAG QA）。\n\n详见纯文本版接口文档：`rag/docs/API_REFERENCE.md`（专为 Agent 自动化调用优化）。",
    "contact": {
      "name": "CustomerServiceAgent"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "健康检查",
        "description": "返回服务状态与各组件（向量库/关键词库/嵌入/精排/生成）当前使用的后端与统计信息。",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/upload": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "上传文档并建立索引",
        "description": "上传单个文档文件，服务端自动解析（.txt/.md/.html/.htm/.json/.pdf）、分块、向量化并写入向量库与关键词库。支持的最大文件大小见服务端配置。",
        "operationId": "upload_document_api_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/ingest_blocks": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "批量导入已分块的文档（如 process/ 输出的 JSON 块）",
        "description": "直接接收结构化文档块数组进行索引，跳过文档解析与分块步骤，适用于对接 process/ 数据处理流水线。",
        "operationId": "ingest_blocks_api_documents_ingest_blocks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBlocksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "知识库文档列表",
        "description": "返回当前知识库中已索引的全部文档及其分块数量。",
        "operationId": "list_documents_api_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{doc_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "文档详情",
        "operationId": "get_document_api_documents__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInfo"
                }
              }
            }
          },
          "404": {
            "description": "文档不存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "删除文档",
        "description": "从向量库、关键词库与知识库登记表中同时删除该文档的全部分块。",
        "operationId": "delete_document_api_documents__doc_id__delete",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "文档不存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{doc_id}/history": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "文档版本历史",
        "description": "返回该文档每次（重新）导入的内容哈希、块数与时间，见 `rag/knowledge_base/versioning.py`。",
        "operationId": "get_document_history_api_documents__doc_id__history_get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {},
                  "title": "Response Get Document History Api Documents  Doc Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/retrieve": {
      "post": {
        "tags": [
          "retrieve"
        ],
        "summary": "检索问答上下文",
        "description": "对用户查询执行「向量检索 + 关键词检索 → 融合去重 → Reranker 精排」全流程，返回精排后的 Top-K 文档块（含来源 page_url、block_path 与相关性分数）。",
        "operationId": "retrieve_api_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "检索增强问答（RAG QA）",
        "description": "对用户问题执行「检索 → 融合去重 → 精排 → 生成」全流程，返回基于知识库上下文生成的回答、引用来源与检索上下文。无生成模型服务时自动降级为抽取式摘录回答。",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "检索增强问答（SSE 流式输出）",
        "description": "与 /chat 相同的问答流程，但以 Server-Sent Events 方式逐块返回答案文本，便于前端流式展示。事件类型：`citations`（先返回引用信息）、`answer`（逐段答案文本）、`done`（结束标记）。",
        "operationId": "chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard": {
      "get": {
        "tags": [
          "observability"
        ],
        "summary": "运维监控看板",
        "description": "返回请求量/延迟分位数/后端使用分布/错误率等监控快照，以及触发的阈值告警列表。",
        "operationId": "dashboard_api_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Dashboard Api Dashboard Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_document_api_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "待上传的文档文件"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_documents_upload_post"
      },
      "ChatRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "用户问题",
            "examples": [
              "千川广告投放规则是什么"
            ]
          },
          "dialogue": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DialogueTurn"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dialogue",
            "description": "多轮历史对话（可选）"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K",
            "description": "检索上下文条数"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "rewritten_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rewritten Query"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationItem"
            },
            "type": "array",
            "title": "Citations"
          },
          "backend_used": {
            "type": "string",
            "title": "Backend Used"
          },
          "contexts": {
            "items": {
              "$ref": "#/components/schemas/ContextItem"
            },
            "type": "array",
            "title": "Contexts"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "query",
          "answer",
          "citations",
          "backend_used",
          "contexts",
          "latency_ms"
        ],
        "title": "ChatResponse"
      },
      "CitationItem": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "page_url": {
            "type": "string",
            "title": "Page Url"
          },
          "block_path": {
            "type": "string",
            "title": "Block Path",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "index",
          "page_url",
          "score"
        ],
        "title": "CitationItem"
      },
      "ContextItem": {
        "properties": {
          "global_chunk_idx": {
            "type": "integer",
            "title": "Global Chunk Idx"
          },
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "chunk_idx": {
            "type": "integer",
            "title": "Chunk Idx"
          },
          "page_name": {
            "type": "string",
            "title": "Page Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "page_url": {
            "type": "string",
            "title": "Page Url"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "html_content": {
            "type": "string",
            "title": "Html Content",
            "default": ""
          },
          "block_path": {
            "type": "string",
            "title": "Block Path",
            "default": ""
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "question": {
            "type": "string",
            "title": "Question",
            "default": ""
          },
          "time": {
            "type": "string",
            "title": "Time",
            "default": ""
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "source_retriever": {
            "type": "string",
            "title": "Source Retriever"
          }
        },
        "type": "object",
        "required": [
          "global_chunk_idx",
          "doc_id",
          "chunk_idx",
          "page_name",
          "title",
          "page_url",
          "text",
          "score",
          "source_retriever"
        ],
        "title": "ContextItem"
      },
      "DeleteResponse": {
        "properties": {
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "deleted"
        ],
        "title": "DeleteResponse"
      },
      "DialogueTurn": {
        "properties": {
          "speaker": {
            "type": "string",
            "title": "Speaker",
            "description": "发言角色：user 或 bot",
            "examples": [
              "user"
            ]
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "发言内容",
            "examples": [
              "广告为什么被限流了？"
            ]
          }
        },
        "type": "object",
        "required": [
          "speaker",
          "text"
        ],
        "title": "DialogueTurn"
      },
      "DocumentInfo": {
        "properties": {
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "num_chunks": {
            "type": "integer",
            "title": "Num Chunks"
          },
          "chunk_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Chunk Ids"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "filename",
          "source",
          "num_chunks",
          "chunk_ids",
          "created_at"
        ],
        "title": "DocumentInfo"
      },
      "DocumentListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentInfo"
            },
            "type": "array",
            "title": "Documents"
          }
        },
        "type": "object",
        "required": [
          "total",
          "documents"
        ],
        "title": "DocumentListResponse"
      },
      "DocumentUploadResponse": {
        "properties": {
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "num_chunks": {
            "type": "integer",
            "title": "Num Chunks"
          },
          "chunk_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Chunk Ids"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "filename",
          "source",
          "num_chunks",
          "chunk_ids",
          "created_at"
        ],
        "title": "DocumentUploadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          }
        },
        "type": "object",
        "required": [
          "status",
          "stats"
        ],
        "title": "HealthResponse"
      },
      "IngestBlocksRequest": {
        "properties": {
          "blocks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Blocks",
            "description": "文档块列表（process/ 输出的 JSON 数组）"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "来源标识（如文件名/页面名）",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "blocks"
        ],
        "title": "IngestBlocksRequest"
      },
      "RetrieveRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "用户查询",
            "examples": [
              "广告限流后怎么办"
            ]
          },
          "dialogue": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DialogueTurn"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dialogue",
            "description": "多轮历史对话，用于 query 重写（可选）"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K",
            "description": "返回条数，默认使用服务端配置"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RetrieveRequest"
      },
      "RetrieveResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ContextItem"
            },
            "type": "array",
            "title": "Results"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "query",
          "results",
          "latency_ms"
        ],
        "title": "RetrieveResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}